[MOV] helpers: move text helpers to a separate file
Task: 3240138
