+++ src/helpers/index.ts
@@ -9,5 +9,6 @@ export * from "./range";
 export * from "./references";
 export * from "./search";
 export * from "./sheet";
+export * from "./text_helper";
 export * from "./uuid";
 export * from "./zones";
